+++ src/Util/HtmlElement.php
@@ -24,7 +24,7 @@ final class HtmlElement implements \Stringable
     /** @var array<string, string|bool> */
     private array $attributes = [];
 
-    /** @var HtmlElement|HtmlElement[]|string */
+    /** @var \Stringable|\Stringable[]|string */
     private $contents;
 
     /** @psalm-readonly */
@@ -33,7 +33,7 @@ final class HtmlElement implements \Stringable
     /**
      * @param string                                $tagName     Name of the HTML tag
      * @param array<string, string|string[]|bool>   $attributes  Array of attributes (values should be unescaped)
-     * @param HtmlElement|HtmlElement[]|string|null $contents    Inner contents, pre-escaped if needed
+     * @param \Stringable|\Stringable[]|string|null $contents    Inner contents, pre-escaped if needed
      * @param bool                                  $selfClosing Whether the tag is self-closing
      */
     public function __construct(string $tagName, array $attributes = [], $contents = '', bool $selfClosing = false)
@@ -89,7 +89,7 @@ final class HtmlElement implements \Stringable
     }
 
     /**
-     * @return HtmlElement|HtmlElement[]|string
+     * @return \Stringable|\Stringable[]|string
      *
      * @psalm-immutable
      */
@@ -105,7 +105,7 @@ final class HtmlElement implements \Stringable
     /**
      * Sets the inner contents of the tag (must be pre-escaped if needed)
      *
-     * @param HtmlElement|HtmlElement[]|string $contents
+     * @param \Stringable|\Stringable[]|string $contents
      *
      * @return $this
      */